requirements.txt
# This file is auto-generated from environment.yml, do not modify.
# See that file for comments about the need/usage of each dependency.

geopandas>=0.12.0
numba==0.*
numpy>=1,<3
matplotlib==3.*
pyproj>=3.4,<4
rasterio>=1.3,<2
scipy==1.*
tqdm
geoutils==0.1.17
affine
pandas
pyogrio
shapely
pip
cerberus
PyYAML
